requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
python-dotenv
